Cpu: try to keep emulated stack below 4GiB
Frogger (https://github.com/xsaco07/Frogger-Uefi-Boot/blob/master/Frogger.asm)
crashes with a corrupted stack because it manipulates ESP (which clears
the high bits of RSP).

Technically that application has way more things wrong with it (it's not
really 64-bit clean). It's be better if there was a way that x64 images
were loaded below 4GiB that if were possible.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>