Cpu: handle SetJump/LongJump from native code
...across emulated code, where from the emulator's stand point
we jump from one native thunk to another, and have stale
contexts to clean up. The simplest example here is
an x64 image starting an x64 image (if we were not handling
gBS->Exit via a wrapper!)

Also detect cases where SetJump comes from native code prior
to any calls into the emulator. The simple example here
is a native image starting an x64 image (if we were not handling
gBS->Exit via a wrapper!). This is not a normal situation and
the checks have overhead, so these are on only by default for
DEBUG builds.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>