Cpu: further tune behavior for periodic timeouts out of uc_emu_start
Reading the TSC even on a physical system is more expensive than
a memory variable.

So the UC_HOOK_BLOCK callback is going to be as simple, fast and short
as possible. mExitPeriodTbs is then re-calibrated once we return
from uc_emu_start.

Now emulated and non-emulated envos can be used with the same build.

Emulated: nearly same performance
Real (Arm): better perf (10% for memory)

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>