Improve logging by adding a SerialPortLib via EFI_SERIAL_IO_PROTOCOL.
This is by far the most versatile mechanism, and should work both
for standalone and firmware direct-included builds on most
systems. The few odd ones without a UART (or with multiple UARTs?)
can use ConOut (but not suggested for firmware builds, it won't
work right for debugging early crashes) or even SBI (RV only).

Another benefit is that we can take out the depex on gEfiSimpleTextOutProtocolGuid.
BaseSerialPortLibSerialIoProtocol does have a depex on gEfiSerialIoProtocolGuid,
and the EmulatorBi depex file is regenerated to include this. This basically
means the emulator driver won't load until the Serial driver is loaded.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>